requirements.txt
fastapi
uvicorn
minio
python-dotenv
pydantic
openai
pinecone-client
PyPDF2
pytesseract
Pillow